+++ webpack.dev.js
@@ -1,4 +1,3 @@
-const { CheckerPlugin } = require('awesome-typescript-loader');
 const merge = require('webpack-merge');
 const common = require('./webpack.common.js');
 
@@ -7,5 +6,4 @@ module.exports = merge(common, {
   // Enable sourcemaps for debugging webpack's output.
   devtool: 'source-map',
   watch: true,
-  plugins: [new CheckerPlugin()],
 });