+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - No longer need to call `cargo.init`. Initialization is done in the constructor and now you only need to call `cargo.enable`
 - Contract ABIs are fetched as needed rather than upfront.
 - Contract ABIs are stored in localstorage to mitigate the need to request them from the server unnecessarily.
+- web3 has been added as a peer dependency
 
 ## [2.2.2]