+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.1]
+
+- Adding support for ERC-1155 purchase and sale
+- Update get resale items method to accept collection address
+
 ## [3.1.0]
 
 - Introducing LAB mehods which are experimental in nature